# apps/web/lib/api/openapi.json
{
  "components": {
    "schemas": {
...
      "GeocodeRequest": {
        "additionalProperties": false,
        "properties": {
          "addresses": {
            "items": {
              "type": "string"
            },
            "maxItems": 60,
            "minItems": 1,
            "title": "Addresses",
            "type": "array"
          }
        },
        "required": [
          "addresses"
        ],
        "title": "GeocodeRequest",
        "type": "object"
      },
      "GeocodeResponse": {
        "description": "Exactly one result per requested address, in the same order.\n\nThe client lines these up against roster rows, so a shorter list would shift every address after\na failure onto the wrong person.",
        "properties": {
          "results": {
            "items": {
              "$ref": "#/components/schemas/ResolutionRead"
            },
            "title": "Results",
            "type": "array"
          }
        },
        "required": [
          "results"
        ],
        "title": "GeocodeResponse",
        "type": "object"
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "PlaceRead": {
        "properties": {
          "address": {
            "title": "Address",
            "type": "string"
          },
          "confidence": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Confidence"
          },
          "is_approximate": {
            "title": "Is Approximate",
            "type": "boolean"
          },
          "lat": {
            "title": "Lat",
            "type": "number"
          },
          "lng": {
            "title": "Lng",
            "type": "number"
          }
        },
        "required": [
          "address",
          "lat",
          "lng",
          "confidence",
          "is_approximate"
        ],
        "title": "PlaceRead",
        "type": "object"
      },
...
      "ResolutionRead": {
        "description": "One result, in the order asked for. `place` is null when nothing matched.",
        "properties": {
          "from_cache": {
            "title": "From Cache",
            "type": "boolean"
          },
          "place": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/PlaceRead"
              },
              {
                "type": "null"
              }
            ]
          },
          "query": {
            "title": "Query",
            "type": "string"
          }
        },
        "required": [
          "query",
          "place",
          "from_cache"
        ],
        "title": "ResolutionRead",
        "type": "object"
      },
...
      "SuggestionsResponse": {
        "properties": {
          "suggestions": {
            "items": {
              "$ref": "#/components/schemas/PlaceRead"
            },
            "title": "Suggestions",
            "type": "array"
          }
        },
        "required": [
          "suggestions"
        ],
        "title": "SuggestionsResponse",
        "type": "object"
      },
...
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      },
...
    },
    "securitySchemes": {
      "HTTPBearer": {
        "description": "Organizer token, shown once at event creation",
        "scheme": "bearer",
        "type": "http"
      }
    }
  },
  "info": {
    "title": "CarpoolOptimizer API",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/v1/geocode": {
      "post": {
        "description": "Resolve a batch of complete addresses, cache-first.\n\nBatched because a pasted roster is the case that matters: forty addresses as forty requests is\nforty round trips and forty cache queries, against a provider limited to 100 calls a minute.",
        "operationId": "geocode_v1_geocode_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/GeocodeRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GeocodeResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Geocode",
        "tags": [
          "geocoding"
        ]
      }
    },
    "/v1/geocode/autocomplete": {
      "get": {
        "description": "Suggestions for a partial address.\n\n**Not cached.** `geocode_cache` is keyed by a full normalized address; a prefix is not one, and\nstoring every keystroke would fill the table with strings no one will ever look up again. The\naddress the coordinator actually picks is resolved through `POST /v1/geocode`, which is cached.",
        "operationId": "autocomplete_v1_geocode_autocomplete_get",
        "parameters": [
          {
            "in": "query",
            "name": "q",
            "required": true,
            "schema": {
              "maxLength": 300,
              "minLength": 3,
              "title": "Q",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SuggestionsResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Autocomplete",
        "tags": [
          "geocoding"
        ]
      }
    }
  }
}